Dodanie widoku wysłanych wiadomości, oraz usuwania wiadomości.

+++ application/models/Message_Model.php
@@ -8,6 +8,7 @@ Class Message_Model extends CI_Model
 public function getmessages($username)
 	{
 		$this->db->where('messto',$username);
+		$this->db->where('delto',0);
 		$s=$this->db->get('messages');
 		return $s;
 	}
@@ -35,15 +36,42 @@ public function getmessages($username)
 		$this->db->where('messto',$username);
 		$this->db->update('messages');
 	}
-	public function delete($id)
+	public function delete($id,$type)
 	{
 		$this->db->where('id',$id);
-		$this->db->delete('messages');
+		if($type=="S")
+		{
+			$this->db->set('delfrom',1);
+		}
+		else if($type=="G")
+		{
+			$this->db->set('delto',1);
+		}
+		$this->db->update('messages');
+		$this->db->where('id',$id);
+		$this->db->select('delto, delfrom');
+		$s=$this->db->get('messages');
+		$row=$s->row(0);
+		if($row->delto)
+		{
+			if($row->delfrom)
+			{
+				$this->db->where('id',$id);
+				$this->db->delete('messages');
+			}
+		}
 	}
 	public function getusers()
 	{
 		$this->db->select('username');
 		$s=$this->db->get('author');
+		return $s;
+	}
+	public function getsended($usr)
+	{
+		$this->db->where('messfrom',$usr);
+		$this->db->where('delfrom',0);
+		$s=$this->db->get('messages');
 		return $s;
 	}
 }
